Add serializer/deserializer
